pyproject.toml
[tool.poetry]
name = "f5-export"
version = "0.1.0"
description = "F5 export tool, exports config data from a F5 BIG-IP"
authors = ["Peter Baumann <peter.baumann@ensec.ch>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "*"
click = "*"
bigrest = "*"
requests = "*"
pre-commit = "*"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[tool.poetry.scripts]
f5export = "f5_export.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
